requirements.txt
pandas
unidecode
fastapi
uvicorn
openpyxl
pymongo
motor
python-dotenv
httpx
# redis
#pip install -r requirements.txt